pyproject.toml
[tool.poetry]
name = "glacier-machine-data-model"
version = "0.0.0"
description = "Machine data model for the Glacier platform."
authors = ["Sebastiano Gaiardelli <sebastiano.gaiardelli@univr.it>"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
unitsnet-py = "^0.1.135"


[tool.poetry.group.dev.dependencies]
radon = "^6.0.1"
lizard = "^1.17.13"
pre-commit = "^4.0.1"
mypy = "^1.13.0"
ruff = "^0.8.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
